Move wizard hook to new import
